Conflicts:
	RootEngine/Render/Source/Renderer.cpp
	RootSystems/Include/CameraSystem.h

+++ RootSystems/Source/RespawnSystem.cpp
@@ -64,11 +64,11 @@ namespace RootSystems
 		for(std::multimap<std::string, ECS::Entity*>::iterator itr = m_spawnPoints.first; itr != m_spawnPoints.second; ++itr, ++numspawns)
 			if(numspawns == chosenspwn)
 			{
-				int x,y,z;
+				float x,y,z;
 				x = m_world->GetEntityManager()->GetComponent<RootForce::Transform>((*itr).second)->m_position.x;
 				y = m_world->GetEntityManager()->GetComponent<RootForce::Transform>((*itr).second)->m_position.y;
 				z = m_world->GetEntityManager()->GetComponent<RootForce::Transform>((*itr).second)->m_position.z;
-				m_engineContext->m_logger->LogText(LogTag::GAME, LogLevel::DEBUG_PRINT, "Found spawnpoint number: %d position: %d %d %d",chosenspwn,x,y,z );
+				m_engineContext->m_logger->LogText(LogTag::GAME, LogLevel::DEBUG_PRINT, "Found spawnpoint number: %d position: %f %f %f",chosenspwn,x,y,z );
 				return m_world->GetEntityManager()->GetComponent<RootForce::Transform>((*itr).second);
 			}
